+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 * **@dlr-eoc/services-map-state**
  - Allow to use `projjson` in `IProjDef`.
 
+* **@dlr-eoc/map-ol:**
+ - Use `projjson` or `proj4js` to create projection if defined on `IProjDef`.
+
+
 # [16.0.0](https://github.com/dlr-eoc/ukis-frontend-libraries/tree/v16.0.0) (2026-05-29) (update dependencies, new library ngx-ukis-ui-clarity and ngx-ukis-utilities, replace core-ui)
 
 ### Breaking Changes